Removed the setup/sources directories.  It doesn't make sense to model a developer in docker.
Made the keystore explicit (${CEDAR_HOME}/.keystore).  JAVA_HOME wasn't working on linux and I suspect that the keystore was not being used.

Made mysql explicily identify itself with a certificate from the dockerized ca.

Made the keystore say that the mysql server was using a certificate from the dockerized ca.

SSL to the database is still not working but it may be a refresh issue.  Try tomorrow from scratch.  Discovered that messaging server eventually switches drivers from one that does not work to one that does work.  Check this more tomorrow.