It has been too long since my last push.  The status is still very similar to what it was.  The call
   java -Djavax.net.ssl.keyStore=/var/CEDAR/.keystore -Djavax.net.ssl.keyStorePassword=changeit \
     -jar target/cedar-admin-tool-2.6.17.jar system-reset

generates an ssl exception

   unable to find valid certification path to requested target

which is pretty clear.  It is surprising though becuase I don't see the problem (unless keycloak is using a localhost certificate which would be bad) and because it is not clear why ssl is being required here.

All the infrastructure and the microservices are running except for the worker microservice that generates the error

          The CedarUser should never be null at this point

The use of certificates and the handling of auth.metadatacenter.orgx have been improved but make no difference on this exception.