Keycloak now lives on keycloak.metadatacenter.orgx which is different from the nginx synonym auth.metadatacenter.orgx.
Rebuilt keycloak to see what went wrong with the realms.

The sleeping to wait for mysql to come up logic is missing.  So to get keycloak to be seen by cedarstatus you need to run it again.  But cedarstatus shows the entire infrastructure.

Finally keycloak waits for mysql with no funky sleep command.

Fixed keycloak.

Removed large files.